+++ docs/examples/react/scroll-fog/preview.tsx
@@ -10,8 +10,8 @@ export default function ScrollFogPreview() {
         borderRadius: "8px",
       }}
     >
-      <ScrollFog>
-        <VStack gap="x4" p="x4" width="full">
+      <ScrollFog placement={["top", "bottom"]}>
+        <VStack gap="x4" px="x4" py="20px" width="full">
           {Array.from({ length: 20 }, (_, i) => (
             <Box key={i} bg="gray" px="x4" py="x3" borderRadius="r2">
               {i + 1}