+++ packages/react/src/components/BottomSheetHandle/BottomSheetHandle.tsx
@@ -1,4 +1,5 @@
 import { Primitive, type PrimitiveProps } from "@seed-design/react-primitive";
+import { Drawer } from "@seed-design/react-drawer";
 import { bottomSheetHandle } from "@seed-design/css/recipes/bottom-sheet-handle";
 import React from "react";
 import clsx from "clsx";
@@ -12,9 +13,9 @@ export const BottomSheetHandle = React.forwardRef<HTMLDivElement, BottomSheetHan
     const classNames = bottomSheetHandle();
 
     return (
-      <Primitive.div ref={ref} className={clsx(classNames.root, className)} {...props}>
+      <Drawer.Handle ref={ref} className={clsx(classNames.root, className)} {...props}>
         <Primitive.div aria-hidden="true" className={classNames.touchArea} />
-      </Primitive.div>
+      </Drawer.Handle>
     );
   },
 );