+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @seed-design/mcp
 
+## 1.1.7
+
+### Patch Changes
+
+- @seed-design/figma@1.1.7
+
+## 1.1.6
+
+### Patch Changes
+
+- @seed-design/figma@1.1.6
+
 ## 1.1.5
 
 ### Patch Changes